+++ visidata/loaders/postgres.py
@@ -71,7 +71,7 @@ class PgTablesSheet(Sheet):
             if r:
                 self.addRow(r)
             self.columns = cursorToColumns(cur)
-            self.addColumn(Column('nrows', type=int, getter=lambda r,self=self: self.getRowCount(r[0])))
+            self.addColumn(Column('nrows', type=int, getter=lambda col,row: col.sheet.getRowCount(row[0])))
 
             for r in cur:
                 self.addRow(r)